Earlier versions required the match_port and match_proto fields to be defined to denote that the plugin should match the port and protocol information. These fields aren't required anymore as the port and protocol information is automatically ommited if it is not found in the list.
The example get_botnets_emerging-botcc script now doesn't support input from stdin anymore. In stead it requires an extra parameter that references to the location of the emerging-bottcc.rules file (which has to be stored on disk). This is because we can then use the timestamp of the file to identify it. The example botnet.cron script reflects this change.

The mail templates belonging to the botnet plugin have been moved from the events-mail plugin to the botnet plugin.




git-svn-id: https://svn.code.sf.net/p/nfsen-plugins/code/nfsen-botnet/trunk@5 b02fbf3d-33ba-4398-9110-6c44321f204a